[llvm/CAS] Move `createOnDiskUnifiedCASDatabases` function to a separate header than `ObjectStore.h`
