+++ llvm/lib/CAS/ObjectStore.cpp
@@ -10,7 +10,6 @@
 #include "BuiltinCAS.h"
 #include "llvm/ADT/DenseSet.h"
 #include "llvm/ADT/StringMap.h"
-#include "llvm/CAS/ActionCache.h"
 #include "llvm/CAS/UnifiedOnDiskCache.h"
 #include "llvm/Support/Debug.h"
 #include "llvm/Support/FileSystem.h"
@@ -191,14 +190,4 @@ cas::createCASFromIdentifier(StringRef Path) {
 void cas::registerCASURLScheme(StringRef Prefix,
                                ObjectStoreCreateFuncTy *Func) {
   getRegisteredScheme().insert({Prefix, Func});
-}
-
-Expected<std::pair<std::unique_ptr<ObjectStore>, std::unique_ptr<ActionCache>>>
-cas::createOnDiskUnifiedCASDatabases(StringRef Path) {
-  std::shared_ptr<ondisk::UnifiedOnDiskCache> UniDB;
-  if (Error E = builtin::createBuiltinUnifiedOnDiskCache(Path).moveInto(UniDB))
-    return std::move(E);
-  auto CAS = builtin::createObjectStoreFromUnifiedOnDiskCache(UniDB);
-  auto AC = builtin::createActionCacheFromUnifiedOnDiskCache(std::move(UniDB));
-  return std::make_pair(std::move(CAS), std::move(AC));
 }
